[Refactor] couchbase service code made more modular and testable, fixed bug 'bucket async method not found
